requirements.txt
arrow==0.15.8
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
defusedxml==0.6.0
Django==2.2.14
django-filter==2.3.0
django-haystack==2.8.1
djangorestframework==3.11.0
docutils==0.16
Faker==4.1.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
loremipsum==1.0.5
MarkupSafe==1.1.1
oauthlib==3.1.0
pathlib==1.0.1
Pillow==7.2.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
python-creole==1.4.6
python-dateutil==2.8.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sh==1.13.1
simplejson==3.17.2
six==1.15.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.10
Whoosh==2.7.4
